refactor: ベンチマークファイルを旧・新で分割
bench_test.go を bench_old_test.go と bench_new_test.go に分割。
